Make importing work more easily
